All data unhooked from view controllers. No view controllers rely on data or crash without. Reorganized folder structures so database is separated from models.
