+++ Pocket Fitness/AppDelegate.swift
@@ -16,7 +16,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
       // Override point for customization after application launch.
-      AppDatabase.setUpSchema()
+      do {
+         try AppDatabase.setUpSchema()
+      } catch {
+         exit(0)
+      }
 
       // TODO: Checkto see if database table exists to preload exercises
