+++ Pocket Fitness/Models/Date.swift
@@ -1,11 +1,3 @@
-//
-//  Date.swift
-//  Pocket Fitness
-//
-//  Created by Vikram Work/School on 5/13/18.
-//  Copyright © 2018 Vikram Work/School. All rights reserved.
-//
-
 import Foundation
 
 extension Date {
