+++ Pocket Fitness/Models/WorkoutExercise.swift
@@ -3,22 +3,22 @@ import SQLite
 
 public class WorkoutExercise {
 
-   var workoutExerciseId : Int64? // Primary key
-   var workoutId : Int64? // Foreign key
-   var exerciseId : Int64? // Foreign key
+   var workoutExerciseId : Int64? 
+   var workoutId : Int64?
+   var exerciseId : Int64?
    var workoutExerciseDate : Date?
-   var userId : Int64? // Foreign key
 
-   // Database
-   
-   
    init(workoutExerciseId : Int64?, workoutId : Int64?,
-        exerciseId : Int64?, workoutExerciseDate : Date?, userId : Int64?) {
+        exerciseId : Int64?, workoutExerciseDate : Date?) {
       self.workoutExerciseId = workoutExerciseId
       self.workoutId = workoutId
       self.exerciseId = exerciseId
       self.workoutExerciseDate = workoutExerciseDate
-      self.userId = userId
+   }
+
+   convenience init()   {
+      self.init(workoutExerciseId: nil, workoutId: nil,
+                exerciseId: nil, workoutExerciseDate: nil)
    }
 }
 
@@ -28,10 +28,4 @@ extension WorkoutExercise : Equatable {
       return lhs.workoutExerciseId == rhs.workoutExerciseId
    }
 
-}
-
-// TODO: Implement all database functions for particular model
-extension WorkoutExercise {
-
-   
 }
